+++ daq_control/src/dataBanks.cc
@@ -33,29 +33,29 @@ void DataBank::addEvent(const Event& event) {
  *
  * @return The serialized data bank as a vector of bytes.
  */
-std::vector<uint8_t> DataBank::serialize() const {
-    std::vector<uint8_t> buffer;
+std::vector<uint32_t> DataBank::serialize() const {
+    std::vector<uint32_t> buffer;
 
     // Write Bank Name (4 bytes)
     buffer.insert(buffer.end(), bankName, bankName + 4);
 
     // Write number of events
     uint32_t eventCount = events.size();
-    buffer.insert(buffer.end(), reinterpret_cast<const uint8_t*>(&eventCount), 
-                  reinterpret_cast<const uint8_t*>(&eventCount) + sizeof(eventCount));
+    buffer.insert(buffer.end(), reinterpret_cast<const uint32_t*>(&eventCount), 
+                  reinterpret_cast<const uint32_t*>(&eventCount) + sizeof(eventCount));
 
     // Serialize each event
     for (const auto& event : events) {
-        buffer.insert(buffer.end(), reinterpret_cast<const uint8_t*>(&event.timestamp),
-                      reinterpret_cast<const uint8_t*>(&event.timestamp) + sizeof(event.timestamp));
+        buffer.insert(buffer.end(), reinterpret_cast<const uint32_t*>(&event.timestamp),
+                      reinterpret_cast<const uint32_t*>(&event.timestamp) + sizeof(event.timestamp));
 
         uint32_t dataSize = event.data.size();
-        buffer.insert(buffer.end(), reinterpret_cast<const uint8_t*>(&dataSize),
-                      reinterpret_cast<const uint8_t*>(&dataSize) + sizeof(dataSize));
+        buffer.insert(buffer.end(), reinterpret_cast<const uint32_t*>(&dataSize),
+                      reinterpret_cast<const uint32_t*>(&dataSize) + sizeof(dataSize));
 
         if (!event.data.empty()) {
-            buffer.insert(buffer.end(), reinterpret_cast<const uint8_t*>(event.data.data()),
-                          reinterpret_cast<const uint8_t*>(event.data.data()) + dataSize * sizeof(uint32_t));
+            buffer.insert(buffer.end(), reinterpret_cast<const uint32_t*>(event.data.data()),
+                          reinterpret_cast<const uint32_t*>(event.data.data()) + dataSize * sizeof(uint32_t));
         }
     }
 
@@ -96,8 +96,8 @@ void Block::addDataBank(const DataBank& bank) {
      * 
      * @return A binary vector containing the serialized block data.
      */
-std::vector<uint8_t> Block::serialize() const {
-    std::vector<uint8_t> buffer;
+std::vector<uint32_t> Block::serialize() const {
+    std::vector<uint32_t> buffer;
 
     // Write Block ID (4 bytes)
     buffer.insert(buffer.end(), reinterpret_cast<const uint8_t*>(&blockID), 
@@ -110,7 +110,7 @@ std::vector<uint8_t> Block::serialize() const {
 
     // Serialize each bank
     for (const auto& bank : banks) {
-        std::vector<uint8_t> bankData = bank.serialize();
+        std::vector<uint32_t> bankData = bank.serialize();
         buffer.insert(buffer.end(), bankData.begin(), bankData.end());
     }
